microservice/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
channels_redis==4.1.0
daphne==4.0.0
redis==5.0.1
pytest-asyncio==0.21.1
pytest-django==4.7.0